requirements.txt
django
djangorestframework
psycopg2-binary
pillow
djangorestframework-simplejwt
djangorestframework-simplejwt[blacklist]
django-cors-headers
django-filter
django-environ